chore: remove logic duplication in 'get_pool' by re-using 'tick_pool'
  Also, relocate tick_pool under the pools::Row implementations, as it
  directly relates to it.
